Renaming oadm_manage_node to oc_adm_manage_node
