+++ roles/lib_openshift/src/ansible/oc_adm_manage_node.py
