+++ roles/lib_openshift/src/class/oc_adm_manage_node.py
